Complete Day 4 part 2.
I actually had a lot of trouble with this one and had to come back to it
the next day. Today I gave myself the time to write more and more test
cases. The test case that finally revealed what was critically wrong
with my code was 111222, which it allowed (as I was just seeing if
multiple matching digits existed), but it's actually not allowed.

In the end, this ended up simplyfying the code that determined the new
criteria to simply "if any matching digits are 2".